+++ src/PlantsPlus/AssemblyInfo.cs
@@ -2,6 +2,6 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("PlantsPlus")]
 [assembly: AssemblyProduct("PlantsPlus")]
-[assembly: AssemblyVersion("1.1.0.102")]
-[assembly: AssemblyFileVersion("1.1.0.102")]
-[assembly: AssemblyInformationalVersion("1.1.0-ml.3")]
+[assembly: AssemblyVersion("1.1.1.13")]
+[assembly: AssemblyFileVersion("1.1.1.13")]
+[assembly: AssemblyInformationalVersion("1.1.1-ml.13")]