+++ src/PlantsPlus/MelonLoaderAssemblyInfo.cs
@@ -1,5 +1,5 @@
 using MelonLoader;
 using PlantsPlus;
 
-[assembly: MelonInfo(typeof(Plugin), "Plants+", "1.1.0-ml.3", "Auro / MelonLoader port")]
+[assembly: MelonInfo(typeof(Plugin), "Plants+", "1.1.1-ml.13", "Auro / MelonLoader port")]
 [assembly: HarmonyDontPatchAll]