+++ src/PlantsPlus/PlantRegister.cs
@@ -1,4 +1,4 @@
-using CustomizeLib.BepInEx;
+using CustomizeLib.MelonLoader;
 using Il2Cpp;
 using PlantsPlus.Plants;
 using System;