+++ src/PlantsPlus/Plants/V11Plants.cs
@@ -1,4 +1,4 @@
-using CustomizeLib.BepInEx;
+using CustomizeLib.MelonLoader;
 using Il2Cpp;
 using Il2CppInterop.Runtime;
 using System;